+++ components/TabPanel/TabPanel.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default function TabPanel(props) {
+  const { children, value, index, ...other } = props;
+
+  return (
+    <div role="tabpanel" hidden={value !== index} {...other}>
+      {children}
+    </div>
+  );
+}